[KQP] Minor refactoring of Tasks Graph (#32820)
