+++ ydb/core/kqp/executer_actor/kqp_scan_executer.cpp
@@ -193,6 +193,7 @@ private:
             const auto& stageInfo = TasksGraph.GetStageInfo(task.StageId);
 
             if (task.Meta.NodeId || stageInfo.Meta.IsSysView()) {
+                // TODO: YQL_ENSURE(task.Meta.Type == TTaskMeta::TTaskType::Scan);
                 // Task with source
                 if (!task.Meta.Reads) {
                     continue;
