+++ ydb/core/kqp/executer_actor/shards_resolver/kqp_shards_resolver_events.h
@@ -4,18 +4,14 @@
 #include <ydb/public/api/protos/ydb_status_codes.pb.h>
 #include <ydb/core/kqp/common/simple/kqp_event_ids.h>
 
-namespace NKikimr::NKqp {
+namespace NKikimr::NKqp::NShardResolver {
 
-namespace NShardResolver {
-    struct TEvShardsResolveStatus : public TEventLocal<TEvShardsResolveStatus,
-        TKqpExecuterEvents::EvShardsResolveStatus>
-    {
-        Ydb::StatusIds::StatusCode Status = Ydb::StatusIds::SUCCESS;
-        NYql::TIssues Issues;
+struct TEvShardsResolveStatus : public TEventLocal<TEvShardsResolveStatus, TKqpExecuterEvents::EvShardsResolveStatus> {
+    Ydb::StatusIds::StatusCode Status = Ydb::StatusIds::SUCCESS;
+    NYql::TIssues Issues;
 
-        TMap<ui64, ui64> ShardNodes;
-        ui32 Unresolved = 0;
-    };
+    TMap<ui64 /* shardId */, ui64 /* nodeId */> ShardsToNodes;
+    ui32 Unresolved = 0;
 };
 
-}
+} // namespace NKikimr::NKqp::NShardResolver
