+++ ydb/core/kqp/ut/olap/kqp_olap_ut.cpp
@@ -2371,7 +2371,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
                 case NKqp::TKqpExecuterEvents::EvShardsResolveStatus: {
 
                     auto* msg = ev->Get<NKqp::NShardResolver::TEvShardsResolveStatus>();
-                    for (auto& [shardId, nodeId]: msg->ShardNodes) {
+                    for (auto& [shardId, nodeId]: msg->ShardsToNodes) {
                         Cerr << "-- nodeId: " << nodeId << Endl;
                         nodeId = runtime->GetNodeId(num);
                         ++num;
@@ -2461,7 +2461,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
                 case NKqp::TKqpExecuterEvents::EvShardsResolveStatus: {
 
                     auto* msg = ev->Get<NKqp::NShardResolver::TEvShardsResolveStatus>();
-                    for (auto& [shardId, nodeId]: msg->ShardNodes) {
+                    for (auto& [shardId, nodeId]: msg->ShardsToNodes) {
                         Cerr << "-- nodeId: " << nodeId << Endl;
                         nodeId = runtime->GetNodeId(0);
                     }
@@ -2524,7 +2524,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
                 case NKqp::TKqpExecuterEvents::EvShardsResolveStatus: {
 
                     auto* msg = ev->Get<NKqp::NShardResolver::TEvShardsResolveStatus>();
-                    for (auto& [shardId, nodeId]: msg->ShardNodes) {
+                    for (auto& [shardId, nodeId]: msg->ShardsToNodes) {
                         Cerr << "-- nodeId: " << nodeId << Endl;
                         nodeId = runtime->GetNodeId(0);
                     }
@@ -2606,7 +2606,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
                 switch (ev->GetTypeRewrite()) {
                     case NKqp::TKqpExecuterEvents::EvShardsResolveStatus: {
                         auto* msg = ev->Get<NKqp::NShardResolver::TEvShardsResolveStatus>();
-                        for (auto& [shardId, nodeId] : msg->ShardNodes) {
+                        for (auto& [shardId, nodeId] : msg->ShardsToNodes) {
                             Cerr << "-- nodeId: " << nodeId << Endl;
                             nodeId = runtime->GetNodeId(0);
                         }
